Function to perform cFDR for continuous auxiliary covariates

Performs Flexible cFDR for continuous auxiliary covariates

flexible_cfdr(
  p,
  q,
  indep_index,
  nxbin = 1000,
  res_p = 300,
  res_q = 500,
  gridp = 50,
  splinecorr = TRUE,
  dist_thr = 0.5,
  plot_KDE = FALSE,
  maf = NULL,
  check_indep_cor = TRUE,
  enforce_p_q_cor = TRUE
)

Arguments

p
p values for principal trait (vector of length n)

q
continuous auxillary data values (vector of length n)

indep_index
indices of independent SNPs

nxbin
number of bins in zp direction for hex-binning

res_p
number of test points in x-direction (p)

res_q
number of test points in y-direction (q)

gridp
number of data points required in a KDE grid point for left-censoring

splinecorr
logical value for whether spline correction should be implemented

dist_thr
distance threshold for spline correction

plot_KDE
logical value for whether to plot the fitted KDE

maf
minor allele frequencies for SNPs to which p and q relate

check_indep_cor
check that sign of the correlation between p and q is the same in the independent subset as in the whole

enforce_p_q_cor
if p and q are negatively correlated, flip the sign on q values

Value

list of length two: (1) dataframe of p-values, q-values and v-values (2) dataframe of auxiliary data (q_low used for left censoring, how many data-points were left censored and/or spline corrected)

Details

If maf is specified, then the independent SNPs will be downsampled to match the minor allele frequency distribution.