parameters_in_locfdr

parameters_in_locfdr

parameters_in_locfdr(
  p,
  q,
  indep_index,
  res_p = 300,
  res_q = 500,
  maf = NULL,
  check_indep_cor = TRUE,
  enforce_p_q_cor = TRUE
)

Arguments

p
p values for principal trait (vector of length n)

q
continuous auxiliary data values (vector of length n)

indep_index
indices of independent SNPs

res_p
resolution for p

res_q
resolution for q

maf
minor allele frequencies for SNPs to which p and q relate (optional and used to perform MAF matching)

check_indep_cor
check that sign of the correlation between p and q is the same in the independent subset as in the whole

enforce_p_q_cor
if p and q are negatively correlated, flip the sign on q values

Value

list of values used as input into locfdr::locfdr function intrinsically in flexible_cfdr

Examples

# In this example, we generate some p-values (representing GWAS p-values)
# and some arbitrary auxiliary data values (e.g. representing functional genomic data).
# We use the parameters_in_locfdr() function to extract the parameters estimated by
# the locfdr function.

# generate p
set.seed(1)
n <- 1000
n1p <- 50 
zp <- c(rnorm(n1p, sd=5), rnorm(n-n1p, sd=1))
p <- 2*pnorm(-abs(zp))

# generate q
mixture_comp1 <- function(x) rnorm(x, mean = -0.5, sd = 0.5)
mixture_comp2 <- function(x) rnorm(x, mean = 2, sd = 1)
q <- c(mixture_comp1(n1p), mixture_comp2(n-n1p))

n_indep <- n

parameters_in_locfdr(p, q, indep_index = 1:n_indep)

#> $zz
#>    [1]   3.132269054   0.918216621   4.178143062   7.976404011   1.647538859
#>    [6]   4.102341921   2.437145262   3.691623526   2.878906758   1.526941936
#>   [11]   7.558905842   1.949216182   3.106202903  11.073499436   5.624654591
#>   [16]   0.224668045   0.080951315   4.719181053   4.106105975   2.969506606
#>   [21]   4.594886858   3.910681504   0.372824917   9.946758479   3.099128739
#>   [26]   0.280643698   0.778977534   7.353761919   2.390750276   2.089707801
#>   [31]   6.793397758   0.513938637   1.938358058   0.269025203   6.885297784
#>   [36]   2.074972816   1.971449769   0.296566984   5.500126860   3.815878742
#>   [41]   0.822617981   1.266808401   3.484816877   2.783315993   3.443778473
#>   [46]   3.537475785   1.822909811   3.842664623   0.561731061   4.405538632
#>   [51]   0.398105880   0.612026393   0.341119691   1.129363096   1.433023702
#>   [56]   1.980399899   0.367221476   1.044134626   0.569719627   0.135054604
#>   [61]   2.401617761   0.039240003   0.689739362   0.028002159   0.743273209
#>   [66]   0.188792300   1.804958629   1.465554862   0.153253338   2.172611670
#>   [71]   0.475509529   0.709946431   0.610726353   0.934097632   1.253633400
#>   [76]   0.291446236   0.443291873   0.001105352   0.074341324   0.589520946
#>   [81]   0.568668733   0.135178615   1.178086997   1.523566800   0.593946188
#>   [86]   0.332950371   1.063099837   0.304183924   0.370018810   0.267098791
#>   [91]   0.542520031   1.207867806   1.160402616   0.700213650   1.586833455
#>   [96]   0.558486426   1.276592208   0.573265414   1.224612615   0.473400636
#>  [101]   0.620366677   0.042115873   0.910921649   0.158028772   0.654584644
#>  [106]   1.767287269   0.716707476   0.910174229   0.384185358   1.682176081
#>  [111]   0.635736454   0.461644730   1.432282239   0.650696353   0.207380744
#>  [116]   0.392807929   0.319992869   0.279113303   0.494188331   0.177330482
#>  [121]   0.505957462   1.343038825   0.214579409   0.179556530   0.100190741
#>  [126]   0.712666307   0.073564404   0.037634171   0.681660479   0.324270272
#>  [131]   0.060160440   0.588894486   0.531496193   1.518394082   0.306557861
#>  [136]   1.536449824   0.300976127   0.528279904   0.652094781   0.056896778
#>  [141]   1.914359426   1.176583312   1.664972436   0.463530401   1.115920105
#>  [146]   0.750819001   2.087166546   0.017395620   1.286300530   1.640605534
#>  [151]   0.450187101   0.018559833   0.318068375   0.929362147   1.487460310
#>  [156]   1.075192297   1.000028804   0.621266695   1.384426847   1.869290622
#>  [161]   0.425100377   0.238647101   1.058483049   0.886422651   0.619243048
#>  [166]   2.206102465   0.255027030   1.424494650   0.144399602   0.207538339
#>  [171]   2.307978399   0.105802368   0.456998805   0.077152935   0.334000842
#>  [176]   0.034726028   0.787639606   2.075245009   1.027392439   1.207908398
#>  [181]   1.231323422   0.983895570   0.219924804   1.467250029   0.521022743
#>  [186]   0.158754605   1.464587312   0.766082000   0.430211754   0.926109497
#>  [191]   0.177103961   0.402011779   0.731748173   0.830373168   1.208082786
#>  [196]   1.047984413   1.441157707   1.015847465   0.411974712   0.381076051
#>  [201]   0.409401840   1.688873286   1.586588433   0.330907801   2.285235535
#>  [206]   2.497661590   0.667066167   0.541327336   0.013399523   0.510108423
#>  [211]   0.164375832   0.420694643   0.400246744   1.370207878   0.987838267
#>  [216]   1.519745025   0.308740569   1.253289756   0.642241306   0.044709137
#>  [221]   1.733218407   0.002131860   0.630300334   0.340968580   1.156572363
#>  [226]   1.803141908   0.331132036   1.605513412   0.197193439   0.263175646
#>  [231]   0.985826700   2.888920672   0.640481703   0.570507636   0.059723276
#>  [236]   0.098178744   0.560820729   1.186458639   1.096777044   0.005344028
#>  [241]   0.707310667   1.034107735   0.223480415   0.878707613   1.162964556
#>  [246]   2.000164945   0.544790740   0.255670709   0.166121037   1.020463909
#>  [251]   0.136221893   0.407167603   0.069654813   0.247664342   0.695550807
#>  [256]   1.146228357   2.403096215   0.572739555   0.374724407   0.425267722
#>  [261]   0.951012808   0.389237182   0.284330662   0.857409778   1.719627299
#>  [266]   0.270054901   0.422184010   1.189113295   0.331032979   0.939829327
#>  [271]   0.258932583   0.394379168   0.851857092   2.649166881   0.156011676
#>  [276]   1.130207267   2.289123980   0.741001157   1.316245160   0.919803678
#>  [281]   0.398130155   0.407528579   1.324258630   0.701231669   0.580614304
#>  [286]   1.001072181   0.668178607   0.945184953   0.433702150   1.005159218
#>  [291]   0.390118664   0.376370292   0.244164924   1.426257342   1.778429287
#>  [296]   0.134447661   0.765598999   0.955136677   0.050565701   0.305815420
#>  [301]   0.893673702   1.047298149   1.971337386   0.383632106   1.654145302
#>  [306]   1.512212694   0.082965734   0.567220915   1.024548480   0.323006503
#>  [311]   1.043612458   0.099078487   0.454136909   0.655781852   0.035922423
#>  [316]   1.069161461   0.483974930   0.121010111   1.294140004   0.494312836
#>  [321]   1.307901520   1.497041009   0.814702731   1.869788790   0.482029504
#>  [326]   0.456135603   0.353400286   0.170489471   0.864035954   0.679230774
#>  [331]   0.327101015   1.569082185   0.367450756   1.364434929   0.334281365
#>  [336]   0.732750042   0.946585640   0.004398704   0.352322306   0.529695509
#>  [341]   0.739589226   1.063457415   0.246210844   0.289499367   2.264889356
#>  [346]   1.408850456   0.916019329   0.191278951   0.803283216   1.887474463
#>  [351]   1.473881181   0.677268492   0.379962687   0.192798426   1.577891795
#>  [356]   0.596234109   1.173576941   0.155642535   1.918909820   0.195258846
#>  [361]   2.592327670   1.314002167   0.635543001   0.429978839   0.169318332
#>  [366]   0.612218174   0.678340177   0.567951972   0.572542604   1.363291256
#>  [371]   0.388722244   0.277914132   0.823081122   0.068840934   1.167662326
#>  [376]   0.008309014   0.128855402   0.145875628   0.163910957   1.763552003
#>  [381]   0.762586512   1.111431081   0.923206953   0.164341838   1.154825187
#>  [386]   0.056521425   2.129360648   0.344845762   1.904955446   0.811170153
#>  [391]   1.324004321   0.615636849   1.091668956   0.306604862   0.110158762
#>  [396]   0.924312773   1.592913754   0.045010598   0.715128401   0.865223100
#>  [401]   1.074440958   1.895654774   0.602997304   0.390867821   0.416222032
#>  [406]   0.375657423   0.366630946   0.295677453   1.441820410   0.697538292
#>  [411]   0.388167506   0.652536452   1.124772447   0.772110803   0.508086216
#>  [416]   0.523620590   1.017754227   0.251164588   1.429993447   1.709121032
#>  [421]   1.435069572   0.710371146   0.065067574   1.759468735   0.569722972
#>  [426]   1.612346798   1.637280647   0.779568513   0.641176934   0.681131394
#>  [431]   2.033285596   0.500963559   1.531798140   0.024997639   0.592984721
#>  [436]   0.198195421   0.892008392   0.025715071   0.647660451   0.646359415
#>  [441]   0.433832740   1.772611185   0.018259711   0.852814994   0.205162903
#>  [446]   3.008048599   1.366111931   0.424102260   0.236803664   2.342723120
#>  [451]   0.961696633   0.604425734   0.752877279   1.555611593   1.453893738
#>  [456]   0.056331836   0.509369407   2.097882960   1.004361979   0.535771722
#>  [461]   0.453037085   2.165368502   1.245746673   0.595498034   0.004884450
#>  [466]   0.279360782   0.705906125   0.628017153   1.480213960   1.083429910
#>  [471]   0.813244257   1.618876849   0.109655699   0.440889371   1.350993980
#>  [476]   1.318609485   0.364384593   0.233499835   1.193955261   0.027909972
#>  [481]   0.357298855   1.146814136   0.517420484   0.362123773   2.350554326
#>  [486]   2.446531376   0.166703279   1.043667439   1.972934934   0.514671633
#>  [491]   1.090573584   2.284659326   0.885617573   0.111106430   3.810276681
#>  [496]   1.108909998   0.307566624   1.106894472   0.347653649   0.873264535
#>  [501]   0.077303123   0.296868642   1.183242240   0.011292688   0.991601036
#>  [506]   1.593967454   1.372711271   0.249610933   1.159424527   1.114222348
#>  [511]   2.528500689   0.935902559   0.967239458   0.047488592   0.403736793
#>  [516]   0.231496128   0.422372408   0.374118395   0.366005775   1.190101447
#>  [521]   0.737327525   0.290666645   0.884849568   0.208006479   0.047730172
#>  [526]   1.684520646   0.144226557   1.180213666   0.681399923   0.143247631
#>  [531]   1.192316444   1.169228653   0.079201709   0.451773753   1.642028213
#>  [536]   0.769592322   0.303360961   1.281737421   0.602222795   0.307022265
#>  [541]   0.418418103   0.355135530   0.513481115   0.018607400   1.318448972
#>  [546]   0.065832000   0.700296078   0.537326132   2.201782322   0.391973744
#>  [551]   0.496960952   0.224874715   1.117143165   0.394994603   1.549830342
#>  [556]   0.743514480   2.331712118   0.812245442   0.501310657   0.510886566
#>  [561]   1.215364041   0.022558628   0.701239300   0.587482026   0.606727941
#>  [566]   1.096640215   0.247509677   0.159901713   0.625778251   0.900434636
#>  [571]   0.994193629   0.849250386   0.805702289   0.467600936   0.848420314
#>  [576]   0.986769864   0.575620289   2.024842045   1.962353191   1.164920931
#>  [581]   1.376519214   0.167679934   1.584629079   1.677888953   0.488296698
#>  [586]   0.878673263   0.144874874   0.468971760   0.376235477   0.761040275
#>  [591]   0.293294934   0.134841264   1.393845816   1.036988690   2.114335148
#>  [596]   0.768278218   0.816160621   0.436106923   0.904705031   0.763086265
#>  [601]   0.341066980   1.502424534   0.528307712   0.542191355   0.136673356
#>  [606]   1.136733853   1.496627154   0.223385644   2.001719228   0.221703816
#>  [611]   0.164372909   0.332623609   0.385207999   1.398754027   2.675740796
#>  [616]   0.423686089   0.298601512   1.792341727   0.248008225   0.247303918
#>  [621]   0.255510379   1.786938100   1.784662816   1.763586348   0.689600222
#>  [626]   1.100740644   0.714509357   0.246470317   0.319786166   1.362644293
#>  [631]   1.227882590   0.511219233   0.731194999   0.019752007   1.572863915
#>  [636]   0.703333270   0.715932089   0.465214906   0.973902306   0.559217730
#>  [641]   2.432639745   0.340484927   0.713033195   0.659037386   0.036402623
#>  [646]   1.593286302   0.847792797   1.850388849   0.323650632   0.255248113
#>  [651]   0.060921227   0.823491629   1.829730485   1.429916216   0.254137143
#>  [656]   2.939773695   0.002415809   0.509665571   1.084720001   0.704832977
#>  [661]   0.330976350   0.976327473   0.843339880   0.970579905   1.771531349
#>  [666]   0.322470342   1.338800742   0.688156028   0.071280652   2.189752359
#>  [671]   1.157707599   1.181688064   0.527368362   1.456628011   0.572967370
#>  [676]   1.433377705   1.055185019   0.733111877   0.210907264   0.998920727
#>  [681]   1.077850323   1.198974383   0.216637035   0.143087030   1.065750091
#>  [686]   0.428623411   0.656179477   0.959394327   1.556052636   1.040796434
#>  [691]   0.930572409   0.075445931   1.967195349   0.755903643   0.461149161
#>  [696]   0.145106631   2.442311321   0.580318685   0.655051998   0.304508837
#>  [701]   0.707568233   1.971572014   0.089998681   0.014017252   1.123456937
#>  [706]   1.344130123   1.523155771   0.421968210   1.360924464   1.753794845
#>  [711]   1.568364734   1.296755570   0.237596251   1.224150136   0.327812680
#>  [716]   2.412450276   0.313792873   1.659878705   0.130953101   1.095888683
#>  [721]   0.489340957   0.778910295   1.743559353   0.078387285   0.975553793
#>  [726]   0.070659825   1.518599529   0.863779033   0.501568385   0.354781330
#>  [731]   0.488428888   0.936293947   1.062408386   0.983820872   0.424247877
#>  [736]   0.451313481   0.925084797   0.198620810   1.194851017   0.495544705
#>  [741]   2.245152574   1.335371360   1.282775206   0.690795904   0.967062668
#>  [746]   1.345793685   1.033665389   0.811776460   1.801725479   1.771541960
#>  [751]   1.454691370   0.845654313   1.250479659   0.667288068   1.290769688
#>  [756]   2.035003538   2.021346994   1.005973488   0.817123601   0.663988284
#>  [761]   0.011281231   0.619677256   1.281238742   0.124261326   0.175741655
#>  [766]   1.692773794   0.642132717   1.282233032   0.140546974   1.112502683
#>  [771]   0.339676972   1.664764639   0.928851762   1.416826835   0.062720778
#>  [776]   0.980902336   1.087150251   0.139327027   0.386272098   1.123585423
#>  [781]   0.759845657   1.148959103   0.842476259   0.391413340   0.891377242
#>  [786]   1.335258713   0.398123478   0.111586803   0.675743917   0.788597935
#>  [791]   0.086986334   1.382284008   0.168490164   0.823190948   0.220894598
#>  [796]   1.029391655   0.010925691   1.224991155   2.596111388   1.169122592
#>  [801]   1.086908817   1.826083013   0.995281807   0.011861781   0.599628395
#>  [806]   0.177947987   0.425981342   0.996658776   0.727660709   1.726630596
#>  [811]   0.353398496   0.726813666   0.668260976   2.424317309   0.235357425
#>  [816]   1.979633321   0.796794539   1.709276181   1.663668712   0.491109552
#>  [821]   0.174055486   0.961290564   0.293826662   0.080999364   0.183661843
#>  [826]   0.166255035   1.269599066   2.349493321   1.412005407   0.016961493
#>  [831]   0.544319353   1.800112333   1.011440176   0.563716556   0.205420795
#>  [836]   1.165461950   2.236322840   0.302265076   1.042506602   0.983542313
#>  [841]   2.005718580   2.070571484   3.055742369   0.261350594   0.454393259
#>  [846]   0.157560555   0.933388728   0.302828276   1.956150222   0.353536709
#>  [851]   0.450424514   0.659550871   1.031420728   2.371022881   0.324576308
#>  [856]   0.944298751   0.765889998   0.953779267   0.398004450   0.311217063
#>  [861]   0.796092713   0.986428344   0.794531662   0.308817972   0.361444766
#>  [866]   1.398791105   0.056070420   1.698873492   0.231852545   0.119090671
#>  [871]   1.772492853   0.343422165   0.623049782   0.439522294   0.505296790
#>  [876]   0.186035137   0.176417798   0.915848207   0.320176726   0.366687297
#>  [881]   0.940611295   0.634702931   0.062488482   0.182837868   1.103641018
#>  [886]   1.752035619   0.953816460   1.644080457   0.866733529   0.266352192
#>  [891]   0.222370495   0.276908510   1.394253041   0.658912000   0.660531281
#>  [896]   0.013255700   0.931480299   1.214689143   2.088340366   0.526152469
#>  [901]   1.541402567   0.194321070   0.264422549   1.118735169   0.650952956
#>  [906]   1.032900239   0.659201499   0.237829400   0.715275947   0.938483039
#>  [911]   0.095354010   0.462819420   1.468882156   0.152686506   1.773762611
#>  [916]   0.648070934   0.199817476   0.689243733   0.036145510   1.943536312
#>  [921]   0.737213734   2.321333933   0.348909346   1.133916661   0.421335268
#>  [926]   0.924556257   1.007062366   0.189474330   0.933916700   0.343910000
#>  [931]   0.814020276   0.915341002   0.171852130   2.402231114   0.795906852
#>  [936]   2.169115869   0.058383499   1.354914450   0.367550766   0.934517579
#>  [941]   0.041639220   0.676112006   0.866436147   0.235175025   0.933970133
#>  [946]   0.813252167   1.348318555   2.251882776   0.493652434   0.474092599
#>  [951]   1.193691666   0.116395472   0.524858556   0.214421481   0.134449535
#>  [956]   0.168509103   0.964732770   0.408778914   0.466438215   2.239782976
#>  [961]   0.795063185   0.019955117   2.514425122   2.210952032   1.488762230
#>  [966]   1.160751876   1.457738225   2.189875949   0.739062681   0.344008769
#>  [971]   0.455876742   0.878110263   0.958463923   0.705844846   2.996949303
#>  [976]   0.961052324   0.380188254   0.505067590   2.027056004   0.064627029
#>  [981]   0.463658648   0.074778333   0.486836236   0.748910821   0.464234581
#>  [986]   0.129420458   0.815460652   0.040150452   0.781416978   0.676452747
#>  [991]   0.482652526   0.669113468   0.512801322   1.048909929   0.121058163
#>  [996]   0.313292885   0.880670730   0.419286891   1.482751678   0.697318199
#> [1001]  -3.132269054  -0.918216621  -4.178143062  -7.976404011  -1.647538859
#> [1006]  -4.102341921  -2.437145262  -3.691623526  -2.878906758  -1.526941936
#> [1011]  -7.558905842  -1.949216182  -3.106202903 -11.073499436  -5.624654591
#> [1016]  -0.224668045  -0.080951315  -4.719181053  -4.106105975  -2.969506606
#> [1021]  -4.594886858  -3.910681504  -0.372824917  -9.946758479  -3.099128739
#> [1026]  -0.280643698  -0.778977534  -7.353761919  -2.390750276  -2.089707801
#> [1031]  -6.793397758  -0.513938637  -1.938358058  -0.269025203  -6.885297784
#> [1036]  -2.074972816  -1.971449769  -0.296566984  -5.500126860  -3.815878742
#> [1041]  -0.822617981  -1.266808401  -3.484816877  -2.783315993  -3.443778473
#> [1046]  -3.537475785  -1.822909811  -3.842664623  -0.561731061  -4.405538632
#> [1051]  -0.398105880  -0.612026393  -0.341119691  -1.129363096  -1.433023702
#> [1056]  -1.980399899  -0.367221476  -1.044134626  -0.569719627  -0.135054604
#> [1061]  -2.401617761  -0.039240003  -0.689739362  -0.028002159  -0.743273209
#> [1066]  -0.188792300  -1.804958629  -1.465554862  -0.153253338  -2.172611670
#> [1071]  -0.475509529  -0.709946431  -0.610726353  -0.934097632  -1.253633400
#> [1076]  -0.291446236  -0.443291873  -0.001105352  -0.074341324  -0.589520946
#> [1081]  -0.568668733  -0.135178615  -1.178086997  -1.523566800  -0.593946188
#> [1086]  -0.332950371  -1.063099837  -0.304183924  -0.370018810  -0.267098791
#> [1091]  -0.542520031  -1.207867806  -1.160402616  -0.700213650  -1.586833455
#> [1096]  -0.558486426  -1.276592208  -0.573265414  -1.224612615  -0.473400636
#> [1101]  -0.620366677  -0.042115873  -0.910921649  -0.158028772  -0.654584644
#> [1106]  -1.767287269  -0.716707476  -0.910174229  -0.384185358  -1.682176081
#> [1111]  -0.635736454  -0.461644730  -1.432282239  -0.650696353  -0.207380744
#> [1116]  -0.392807929  -0.319992869  -0.279113303  -0.494188331  -0.177330482
#> [1121]  -0.505957462  -1.343038825  -0.214579409  -0.179556530  -0.100190741
#> [1126]  -0.712666307  -0.073564404  -0.037634171  -0.681660479  -0.324270272
#> [1131]  -0.060160440  -0.588894486  -0.531496193  -1.518394082  -0.306557861
#> [1136]  -1.536449824  -0.300976127  -0.528279904  -0.652094781  -0.056896778
#> [1141]  -1.914359426  -1.176583312  -1.664972436  -0.463530401  -1.115920105
#> [1146]  -0.750819001  -2.087166546  -0.017395620  -1.286300530  -1.640605534
#> [1151]  -0.450187101  -0.018559833  -0.318068375  -0.929362147  -1.487460310
#> [1156]  -1.075192297  -1.000028804  -0.621266695  -1.384426847  -1.869290622
#> [1161]  -0.425100377  -0.238647101  -1.058483049  -0.886422651  -0.619243048
#> [1166]  -2.206102465  -0.255027030  -1.424494650  -0.144399602  -0.207538339
#> [1171]  -2.307978399  -0.105802368  -0.456998805  -0.077152935  -0.334000842
#> [1176]  -0.034726028  -0.787639606  -2.075245009  -1.027392439  -1.207908398
#> [1181]  -1.231323422  -0.983895570  -0.219924804  -1.467250029  -0.521022743
#> [1186]  -0.158754605  -1.464587312  -0.766082000  -0.430211754  -0.926109497
#> [1191]  -0.177103961  -0.402011779  -0.731748173  -0.830373168  -1.208082786
#> [1196]  -1.047984413  -1.441157707  -1.015847465  -0.411974712  -0.381076051
#> [1201]  -0.409401840  -1.688873286  -1.586588433  -0.330907801  -2.285235535
#> [1206]  -2.497661590  -0.667066167  -0.541327336  -0.013399523  -0.510108423
#> [1211]  -0.164375832  -0.420694643  -0.400246744  -1.370207878  -0.987838267
#> [1216]  -1.519745025  -0.308740569  -1.253289756  -0.642241306  -0.044709137
#> [1221]  -1.733218407  -0.002131860  -0.630300334  -0.340968580  -1.156572363
#> [1226]  -1.803141908  -0.331132036  -1.605513412  -0.197193439  -0.263175646
#> [1231]  -0.985826700  -2.888920672  -0.640481703  -0.570507636  -0.059723276
#> [1236]  -0.098178744  -0.560820729  -1.186458639  -1.096777044  -0.005344028
#> [1241]  -0.707310667  -1.034107735  -0.223480415  -0.878707613  -1.162964556
#> [1246]  -2.000164945  -0.544790740  -0.255670709  -0.166121037  -1.020463909
#> [1251]  -0.136221893  -0.407167603  -0.069654813  -0.247664342  -0.695550807
#> [1256]  -1.146228357  -2.403096215  -0.572739555  -0.374724407  -0.425267722
#> [1261]  -0.951012808  -0.389237182  -0.284330662  -0.857409778  -1.719627299
#> [1266]  -0.270054901  -0.422184010  -1.189113295  -0.331032979  -0.939829327
#> [1271]  -0.258932583  -0.394379168  -0.851857092  -2.649166881  -0.156011676
#> [1276]  -1.130207267  -2.289123980  -0.741001157  -1.316245160  -0.919803678
#> [1281]  -0.398130155  -0.407528579  -1.324258630  -0.701231669  -0.580614304
#> [1286]  -1.001072181  -0.668178607  -0.945184953  -0.433702150  -1.005159218
#> [1291]  -0.390118664  -0.376370292  -0.244164924  -1.426257342  -1.778429287
#> [1296]  -0.134447661  -0.765598999  -0.955136677  -0.050565701  -0.305815420
#> [1301]  -0.893673702  -1.047298149  -1.971337386  -0.383632106  -1.654145302
#> [1306]  -1.512212694  -0.082965734  -0.567220915  -1.024548480  -0.323006503
#> [1311]  -1.043612458  -0.099078487  -0.454136909  -0.655781852  -0.035922423
#> [1316]  -1.069161461  -0.483974930  -0.121010111  -1.294140004  -0.494312836
#> [1321]  -1.307901520  -1.497041009  -0.814702731  -1.869788790  -0.482029504
#> [1326]  -0.456135603  -0.353400286  -0.170489471  -0.864035954  -0.679230774
#> [1331]  -0.327101015  -1.569082185  -0.367450756  -1.364434929  -0.334281365
#> [1336]  -0.732750042  -0.946585640  -0.004398704  -0.352322306  -0.529695509
#> [1341]  -0.739589226  -1.063457415  -0.246210844  -0.289499367  -2.264889356
#> [1346]  -1.408850456  -0.916019329  -0.191278951  -0.803283216  -1.887474463
#> [1351]  -1.473881181  -0.677268492  -0.379962687  -0.192798426  -1.577891795
#> [1356]  -0.596234109  -1.173576941  -0.155642535  -1.918909820  -0.195258846
#> [1361]  -2.592327670  -1.314002167  -0.635543001  -0.429978839  -0.169318332
#> [1366]  -0.612218174  -0.678340177  -0.567951972  -0.572542604  -1.363291256
#> [1371]  -0.388722244  -0.277914132  -0.823081122  -0.068840934  -1.167662326
#> [1376]  -0.008309014  -0.128855402  -0.145875628  -0.163910957  -1.763552003
#> [1381]  -0.762586512  -1.111431081  -0.923206953  -0.164341838  -1.154825187
#> [1386]  -0.056521425  -2.129360648  -0.344845762  -1.904955446  -0.811170153
#> [1391]  -1.324004321  -0.615636849  -1.091668956  -0.306604862  -0.110158762
#> [1396]  -0.924312773  -1.592913754  -0.045010598  -0.715128401  -0.865223100
#> [1401]  -1.074440958  -1.895654774  -0.602997304  -0.390867821  -0.416222032
#> [1406]  -0.375657423  -0.366630946  -0.295677453  -1.441820410  -0.697538292
#> [1411]  -0.388167506  -0.652536452  -1.124772447  -0.772110803  -0.508086216
#> [1416]  -0.523620590  -1.017754227  -0.251164588  -1.429993447  -1.709121032
#> [1421]  -1.435069572  -0.710371146  -0.065067574  -1.759468735  -0.569722972
#> [1426]  -1.612346798  -1.637280647  -0.779568513  -0.641176934  -0.681131394
#> [1431]  -2.033285596  -0.500963559  -1.531798140  -0.024997639  -0.592984721
#> [1436]  -0.198195421  -0.892008392  -0.025715071  -0.647660451  -0.646359415
#> [1441]  -0.433832740  -1.772611185  -0.018259711  -0.852814994  -0.205162903
#> [1446]  -3.008048599  -1.366111931  -0.424102260  -0.236803664  -2.342723120
#> [1451]  -0.961696633  -0.604425734  -0.752877279  -1.555611593  -1.453893738
#> [1456]  -0.056331836  -0.509369407  -2.097882960  -1.004361979  -0.535771722
#> [1461]  -0.453037085  -2.165368502  -1.245746673  -0.595498034  -0.004884450
#> [1466]  -0.279360782  -0.705906125  -0.628017153  -1.480213960  -1.083429910
#> [1471]  -0.813244257  -1.618876849  -0.109655699  -0.440889371  -1.350993980
#> [1476]  -1.318609485  -0.364384593  -0.233499835  -1.193955261  -0.027909972
#> [1481]  -0.357298855  -1.146814136  -0.517420484  -0.362123773  -2.350554326
#> [1486]  -2.446531376  -0.166703279  -1.043667439  -1.972934934  -0.514671633
#> [1491]  -1.090573584  -2.284659326  -0.885617573  -0.111106430  -3.810276681
#> [1496]  -1.108909998  -0.307566624  -1.106894472  -0.347653649  -0.873264535
#> [1501]  -0.077303123  -0.296868642  -1.183242240  -0.011292688  -0.991601036
#> [1506]  -1.593967454  -1.372711271  -0.249610933  -1.159424527  -1.114222348
#> [1511]  -2.528500689  -0.935902559  -0.967239458  -0.047488592  -0.403736793
#> [1516]  -0.231496128  -0.422372408  -0.374118395  -0.366005775  -1.190101447
#> [1521]  -0.737327525  -0.290666645  -0.884849568  -0.208006479  -0.047730172
#> [1526]  -1.684520646  -0.144226557  -1.180213666  -0.681399923  -0.143247631
#> [1531]  -1.192316444  -1.169228653  -0.079201709  -0.451773753  -1.642028213
#> [1536]  -0.769592322  -0.303360961  -1.281737421  -0.602222795  -0.307022265
#> [1541]  -0.418418103  -0.355135530  -0.513481115  -0.018607400  -1.318448972
#> [1546]  -0.065832000  -0.700296078  -0.537326132  -2.201782322  -0.391973744
#> [1551]  -0.496960952  -0.224874715  -1.117143165  -0.394994603  -1.549830342
#> [1556]  -0.743514480  -2.331712118  -0.812245442  -0.501310657  -0.510886566
#> [1561]  -1.215364041  -0.022558628  -0.701239300  -0.587482026  -0.606727941
#> [1566]  -1.096640215  -0.247509677  -0.159901713  -0.625778251  -0.900434636
#> [1571]  -0.994193629  -0.849250386  -0.805702289  -0.467600936  -0.848420314
#> [1576]  -0.986769864  -0.575620289  -2.024842045  -1.962353191  -1.164920931
#> [1581]  -1.376519214  -0.167679934  -1.584629079  -1.677888953  -0.488296698
#> [1586]  -0.878673263  -0.144874874  -0.468971760  -0.376235477  -0.761040275
#> [1591]  -0.293294934  -0.134841264  -1.393845816  -1.036988690  -2.114335148
#> [1596]  -0.768278218  -0.816160621  -0.436106923  -0.904705031  -0.763086265
#> [1601]  -0.341066980  -1.502424534  -0.528307712  -0.542191355  -0.136673356
#> [1606]  -1.136733853  -1.496627154  -0.223385644  -2.001719228  -0.221703816
#> [1611]  -0.164372909  -0.332623609  -0.385207999  -1.398754027  -2.675740796
#> [1616]  -0.423686089  -0.298601512  -1.792341727  -0.248008225  -0.247303918
#> [1621]  -0.255510379  -1.786938100  -1.784662816  -1.763586348  -0.689600222
#> [1626]  -1.100740644  -0.714509357  -0.246470317  -0.319786166  -1.362644293
#> [1631]  -1.227882590  -0.511219233  -0.731194999  -0.019752007  -1.572863915
#> [1636]  -0.703333270  -0.715932089  -0.465214906  -0.973902306  -0.559217730
#> [1641]  -2.432639745  -0.340484927  -0.713033195  -0.659037386  -0.036402623
#> [1646]  -1.593286302  -0.847792797  -1.850388849  -0.323650632  -0.255248113
#> [1651]  -0.060921227  -0.823491629  -1.829730485  -1.429916216  -0.254137143
#> [1656]  -2.939773695  -0.002415809  -0.509665571  -1.084720001  -0.704832977
#> [1661]  -0.330976350  -0.976327473  -0.843339880  -0.970579905  -1.771531349
#> [1666]  -0.322470342  -1.338800742  -0.688156028  -0.071280652  -2.189752359
#> [1671]  -1.157707599  -1.181688064  -0.527368362  -1.456628011  -0.572967370
#> [1676]  -1.433377705  -1.055185019  -0.733111877  -0.210907264  -0.998920727
#> [1681]  -1.077850323  -1.198974383  -0.216637035  -0.143087030  -1.065750091
#> [1686]  -0.428623411  -0.656179477  -0.959394327  -1.556052636  -1.040796434
#> [1691]  -0.930572409  -0.075445931  -1.967195349  -0.755903643  -0.461149161
#> [1696]  -0.145106631  -2.442311321  -0.580318685  -0.655051998  -0.304508837
#> [1701]  -0.707568233  -1.971572014  -0.089998681  -0.014017252  -1.123456937
#> [1706]  -1.344130123  -1.523155771  -0.421968210  -1.360924464  -1.753794845
#> [1711]  -1.568364734  -1.296755570  -0.237596251  -1.224150136  -0.327812680
#> [1716]  -2.412450276  -0.313792873  -1.659878705  -0.130953101  -1.095888683
#> [1721]  -0.489340957  -0.778910295  -1.743559353  -0.078387285  -0.975553793
#> [1726]  -0.070659825  -1.518599529  -0.863779033  -0.501568385  -0.354781330
#> [1731]  -0.488428888  -0.936293947  -1.062408386  -0.983820872  -0.424247877
#> [1736]  -0.451313481  -0.925084797  -0.198620810  -1.194851017  -0.495544705
#> [1741]  -2.245152574  -1.335371360  -1.282775206  -0.690795904  -0.967062668
#> [1746]  -1.345793685  -1.033665389  -0.811776460  -1.801725479  -1.771541960
#> [1751]  -1.454691370  -0.845654313  -1.250479659  -0.667288068  -1.290769688
#> [1756]  -2.035003538  -2.021346994  -1.005973488  -0.817123601  -0.663988284
#> [1761]  -0.011281231  -0.619677256  -1.281238742  -0.124261326  -0.175741655
#> [1766]  -1.692773794  -0.642132717  -1.282233032  -0.140546974  -1.112502683
#> [1771]  -0.339676972  -1.664764639  -0.928851762  -1.416826835  -0.062720778
#> [1776]  -0.980902336  -1.087150251  -0.139327027  -0.386272098  -1.123585423
#> [1781]  -0.759845657  -1.148959103  -0.842476259  -0.391413340  -0.891377242
#> [1786]  -1.335258713  -0.398123478  -0.111586803  -0.675743917  -0.788597935
#> [1791]  -0.086986334  -1.382284008  -0.168490164  -0.823190948  -0.220894598
#> [1796]  -1.029391655  -0.010925691  -1.224991155  -2.596111388  -1.169122592
#> [1801]  -1.086908817  -1.826083013  -0.995281807  -0.011861781  -0.599628395
#> [1806]  -0.177947987  -0.425981342  -0.996658776  -0.727660709  -1.726630596
#> [1811]  -0.353398496  -0.726813666  -0.668260976  -2.424317309  -0.235357425
#> [1816]  -1.979633321  -0.796794539  -1.709276181  -1.663668712  -0.491109552
#> [1821]  -0.174055486  -0.961290564  -0.293826662  -0.080999364  -0.183661843
#> [1826]  -0.166255035  -1.269599066  -2.349493321  -1.412005407  -0.016961493
#> [1831]  -0.544319353  -1.800112333  -1.011440176  -0.563716556  -0.205420795
#> [1836]  -1.165461950  -2.236322840  -0.302265076  -1.042506602  -0.983542313
#> [1841]  -2.005718580  -2.070571484  -3.055742369  -0.261350594  -0.454393259
#> [1846]  -0.157560555  -0.933388728  -0.302828276  -1.956150222  -0.353536709
#> [1851]  -0.450424514  -0.659550871  -1.031420728  -2.371022881  -0.324576308
#> [1856]  -0.944298751  -0.765889998  -0.953779267  -0.398004450  -0.311217063
#> [1861]  -0.796092713  -0.986428344  -0.794531662  -0.308817972  -0.361444766
#> [1866]  -1.398791105  -0.056070420  -1.698873492  -0.231852545  -0.119090671
#> [1871]  -1.772492853  -0.343422165  -0.623049782  -0.439522294  -0.505296790
#> [1876]  -0.186035137  -0.176417798  -0.915848207  -0.320176726  -0.366687297
#> [1881]  -0.940611295  -0.634702931  -0.062488482  -0.182837868  -1.103641018
#> [1886]  -1.752035619  -0.953816460  -1.644080457  -0.866733529  -0.266352192
#> [1891]  -0.222370495  -0.276908510  -1.394253041  -0.658912000  -0.660531281
#> [1896]  -0.013255700  -0.931480299  -1.214689143  -2.088340366  -0.526152469
#> [1901]  -1.541402567  -0.194321070  -0.264422549  -1.118735169  -0.650952956
#> [1906]  -1.032900239  -0.659201499  -0.237829400  -0.715275947  -0.938483039
#> [1911]  -0.095354010  -0.462819420  -1.468882156  -0.152686506  -1.773762611
#> [1916]  -0.648070934  -0.199817476  -0.689243733  -0.036145510  -1.943536312
#> [1921]  -0.737213734  -2.321333933  -0.348909346  -1.133916661  -0.421335268
#> [1926]  -0.924556257  -1.007062366  -0.189474330  -0.933916700  -0.343910000
#> [1931]  -0.814020276  -0.915341002  -0.171852130  -2.402231114  -0.795906852
#> [1936]  -2.169115869  -0.058383499  -1.354914450  -0.367550766  -0.934517579
#> [1941]  -0.041639220  -0.676112006  -0.866436147  -0.235175025  -0.933970133
#> [1946]  -0.813252167  -1.348318555  -2.251882776  -0.493652434  -0.474092599
#> [1951]  -1.193691666  -0.116395472  -0.524858556  -0.214421481  -0.134449535
#> [1956]  -0.168509103  -0.964732770  -0.408778914  -0.466438215  -2.239782976
#> [1961]  -0.795063185  -0.019955117  -2.514425122  -2.210952032  -1.488762230
#> [1966]  -1.160751876  -1.457738225  -2.189875949  -0.739062681  -0.344008769
#> [1971]  -0.455876742  -0.878110263  -0.958463923  -0.705844846  -2.996949303
#> [1976]  -0.961052324  -0.380188254  -0.505067590  -2.027056004  -0.064627029
#> [1981]  -0.463658648  -0.074778333  -0.486836236  -0.748910821  -0.464234581
#> [1986]  -0.129420458  -0.815460652  -0.040150452  -0.781416978  -0.676452747
#> [1991]  -0.482652526  -0.669113468  -0.512801322  -1.048909929  -0.121058163
#> [1996]  -0.313292885  -0.880670730  -0.419286891  -1.482751678  -0.697318199
#> 
#> $bre
#>   [1]   0.02036913   0.06110739   0.10184564   0.14258390   0.18332216
#>   [6]   0.22406041   0.26479867   0.30553693   0.34627518   0.38701344
#>  [11]   0.42775170   0.46848996   0.50922821   0.54996647   0.59070473
#>  [16]   0.63144298   0.67218124   0.71291950   0.75365775   0.79439601
#>  [21]   0.83513427   0.87587253   0.91661078   0.95734904   0.99808730
#>  [26]   1.03882555   1.07956381   1.12030207   1.16104032   1.20177858
#>  [31]   1.24251684   1.28325510   1.32399335   1.36473161   1.40546987
#>  [36]   1.44620812   1.48694638   1.52768464   1.56842289   1.60916115
#>  [41]   1.64989941   1.69063767   1.73137592   1.77211418   1.81285244
#>  [46]   1.85359069   1.89432895   1.93506721   1.97580546   2.01654372
#>  [51]   2.05728198   2.09802024   2.13875849   2.17949675   2.22023501
#>  [56]   2.26097326   2.30171152   2.34244978   2.38318803   2.42392629
#>  [61]   2.46466455   2.50540281   2.54614106   2.58687932   2.62761758
#>  [66]   2.66835583   2.70909409   2.74983235   2.79057060   2.83130886
#>  [71]   2.87204712   2.91278538   2.95352363   2.99426189   3.03500015
#>  [76]   3.07573840   3.11647666   3.15721492   3.19795317   3.23869143
#>  [81]   3.27942969   3.32016795   3.36090620   3.40164446   3.44238272
#>  [86]   3.48312097   3.52385923   3.56459749   3.60533574   3.64607400
#>  [91]   3.68681226   3.72755052   3.76828877   3.80902703   3.84976529
#>  [96]   3.89050354   3.93124180   3.97198006   4.01271831   4.05345657
#> [101]   4.09419483   4.13493309   4.17567134   4.21640960   4.25714786
#> [106]   4.29788611   4.33862437   4.37936263   4.42010088   4.46083914
#> [111]   4.50157740   4.54231566   4.58305391   4.62379217   4.66453043
#> [116]   4.70526868   4.74600694   4.78674520   4.82748346   4.86822171
#> [121]   4.90895997   4.94969823   4.99043648   5.03117474   5.07191300
#> [126]   5.11265125   5.15338951   5.19412777   5.23486603   5.27560428
#> [131]   5.31634254   5.35708080   5.39781905   5.43855731   5.47929557
#> [136]   5.52003382   5.56077208   5.60151034   5.64224860   5.68298685
#> [141]   5.72372511   5.76446337   5.80520162   5.84593988   5.88667814
#> [146]   5.92741639   5.96815465   6.00889291   6.04963117   6.09036942
#> [151]   6.13110768   6.17184594   6.21258419   6.25332245   6.29406071
#> [156]   6.33479896   6.37553722   6.41627548   6.45701374   6.49775199
#> [161]   6.53849025   6.57922851   6.61996676   6.66070502   6.70144328
#> [166]   6.74218153   6.78291979   6.82365805   6.86439631   6.90513456
#> [171]   6.94587282   6.98661108   7.02734933   7.06808759   7.10882585
#> [176]   7.14956410   7.19030236   7.23104062   7.27177888   7.31251713
#> [181]   7.35325539   7.39399365   7.43473190   7.47547016   7.51620842
#> [186]   7.55694667   7.59768493   7.63842319   7.67916145   7.71989970
#> [191]   7.76063796   7.80137622   7.84211447   7.88285273   7.92359099
#> [196]   7.96432924   8.00506750   8.04580576   8.08654402   8.12728227
#> [201]   8.16802053   8.20875879   8.24949704   8.29023530   8.33097356
#> [206]   8.37171181   8.41245007   8.45318833   8.49392659   8.53466484
#> [211]   8.57540310   8.61614136   8.65687961   8.69761787   8.73835613
#> [216]   8.77909438   8.81983264   8.86057090   8.90130916   8.94204741
#> [221]   8.98278567   9.02352393   9.06426218   9.10500044   9.14573870
#> [226]   9.18647695   9.22721521   9.26795347   9.30869173   9.34942998
#> [231]   9.39016824   9.43090650   9.47164475   9.51238301   9.55312127
#> [236]   9.59385952   9.63459778   9.67533604   9.71607430   9.75681255
#> [241]   9.79755081   9.83828907   9.87902732   9.91976558   9.96050384
#> [246]  10.00124209  10.04198035  10.08271861  10.12345687  10.16419512
#> [251]  10.20493338  10.24567164  10.28640989  10.32714815  10.36788641
#> [256]  10.40862466  10.44936292  10.49010118  10.53083944  10.57157769
#> [261]  10.61231595  10.65305421  10.69379246  10.73453072  10.77526898
#> [266]  10.81600723  10.85674549  10.89748375  10.93822201  10.97896026
#> [271]  11.01969852  11.06043678  11.10117503  11.14191329  11.18265155
#> [276]  11.22338980  11.26412806  11.30486632  11.34560458  11.38634283
#> [281]  11.42708109  11.46781935  11.50855760  11.54929586  11.59003412
#> [286]  11.63077237  11.67151063  11.71224889  11.75298715  11.79372540
#> [291]  11.83446366  11.87520192  11.91594017  11.95667843  11.99741669
#> [296]  12.03815494  12.07889320  12.11963146  12.16036972  12.20110797
#> [301]  -0.02036913  -0.06110739  -0.10184564  -0.14258390  -0.18332216
#> [306]  -0.22406041  -0.26479867  -0.30553693  -0.34627518  -0.38701344
#> [311]  -0.42775170  -0.46848996  -0.50922821  -0.54996647  -0.59070473
#> [316]  -0.63144298  -0.67218124  -0.71291950  -0.75365775  -0.79439601
#> [321]  -0.83513427  -0.87587253  -0.91661078  -0.95734904  -0.99808730
#> [326]  -1.03882555  -1.07956381  -1.12030207  -1.16104032  -1.20177858
#> [331]  -1.24251684  -1.28325510  -1.32399335  -1.36473161  -1.40546987
#> [336]  -1.44620812  -1.48694638  -1.52768464  -1.56842289  -1.60916115
#> [341]  -1.64989941  -1.69063767  -1.73137592  -1.77211418  -1.81285244
#> [346]  -1.85359069  -1.89432895  -1.93506721  -1.97580546  -2.01654372
#> [351]  -2.05728198  -2.09802024  -2.13875849  -2.17949675  -2.22023501
#> [356]  -2.26097326  -2.30171152  -2.34244978  -2.38318803  -2.42392629
#> [361]  -2.46466455  -2.50540281  -2.54614106  -2.58687932  -2.62761758
#> [366]  -2.66835583  -2.70909409  -2.74983235  -2.79057060  -2.83130886
#> [371]  -2.87204712  -2.91278538  -2.95352363  -2.99426189  -3.03500015
#> [376]  -3.07573840  -3.11647666  -3.15721492  -3.19795317  -3.23869143
#> [381]  -3.27942969  -3.32016795  -3.36090620  -3.40164446  -3.44238272
#> [386]  -3.48312097  -3.52385923  -3.56459749  -3.60533574  -3.64607400
#> [391]  -3.68681226  -3.72755052  -3.76828877  -3.80902703  -3.84976529
#> [396]  -3.89050354  -3.93124180  -3.97198006  -4.01271831  -4.05345657
#> [401]  -4.09419483  -4.13493309  -4.17567134  -4.21640960  -4.25714786
#> [406]  -4.29788611  -4.33862437  -4.37936263  -4.42010088  -4.46083914
#> [411]  -4.50157740  -4.54231566  -4.58305391  -4.62379217  -4.66453043
#> [416]  -4.70526868  -4.74600694  -4.78674520  -4.82748346  -4.86822171
#> [421]  -4.90895997  -4.94969823  -4.99043648  -5.03117474  -5.07191300
#> [426]  -5.11265125  -5.15338951  -5.19412777  -5.23486603  -5.27560428
#> [431]  -5.31634254  -5.35708080  -5.39781905  -5.43855731  -5.47929557
#> [436]  -5.52003382  -5.56077208  -5.60151034  -5.64224860  -5.68298685
#> [441]  -5.72372511  -5.76446337  -5.80520162  -5.84593988  -5.88667814
#> [446]  -5.92741639  -5.96815465  -6.00889291  -6.04963117  -6.09036942
#> [451]  -6.13110768  -6.17184594  -6.21258419  -6.25332245  -6.29406071
#> [456]  -6.33479896  -6.37553722  -6.41627548  -6.45701374  -6.49775199
#> [461]  -6.53849025  -6.57922851  -6.61996676  -6.66070502  -6.70144328
#> [466]  -6.74218153  -6.78291979  -6.82365805  -6.86439631  -6.90513456
#> [471]  -6.94587282  -6.98661108  -7.02734933  -7.06808759  -7.10882585
#> [476]  -7.14956410  -7.19030236  -7.23104062  -7.27177888  -7.31251713
#> [481]  -7.35325539  -7.39399365  -7.43473190  -7.47547016  -7.51620842
#> [486]  -7.55694667  -7.59768493  -7.63842319  -7.67916145  -7.71989970
#> [491]  -7.76063796  -7.80137622  -7.84211447  -7.88285273  -7.92359099
#> [496]  -7.96432924  -8.00506750  -8.04580576  -8.08654402  -8.12728227
#> [501]  -8.16802053  -8.20875879  -8.24949704  -8.29023530  -8.33097356
#> [506]  -8.37171181  -8.41245007  -8.45318833  -8.49392659  -8.53466484
#> [511]  -8.57540310  -8.61614136  -8.65687961  -8.69761787  -8.73835613
#> [516]  -8.77909438  -8.81983264  -8.86057090  -8.90130916  -8.94204741
#> [521]  -8.98278567  -9.02352393  -9.06426218  -9.10500044  -9.14573870
#> [526]  -9.18647695  -9.22721521  -9.26795347  -9.30869173  -9.34942998
#> [531]  -9.39016824  -9.43090650  -9.47164475  -9.51238301  -9.55312127
#> [536]  -9.59385952  -9.63459778  -9.67533604  -9.71607430  -9.75681255
#> [541]  -9.79755081  -9.83828907  -9.87902732  -9.91976558  -9.96050384
#> [546] -10.00124209 -10.04198035 -10.08271861 -10.12345687 -10.16419512
#> [551] -10.20493338 -10.24567164 -10.28640989 -10.32714815 -10.36788641
#> [556] -10.40862466 -10.44936292 -10.49010118 -10.53083944 -10.57157769
#> [561] -10.61231595 -10.65305421 -10.69379246 -10.73453072 -10.77526898
#> [566] -10.81600723 -10.85674549 -10.89748375 -10.93822201 -10.97896026
#> [571] -11.01969852 -11.06043678 -11.10117503 -11.14191329 -11.18265155
#> [576] -11.22338980 -11.26412806 -11.30486632 -11.34560458 -11.38634283
#> [581] -11.42708109 -11.46781935 -11.50855760 -11.54929586 -11.59003412
#> [586] -11.63077237 -11.67151063 -11.71224889 -11.75298715 -11.79372540
#> [591] -11.83446366 -11.87520192 -11.91594017 -11.95667843 -11.99741669
#> [596] -12.03815494 -12.07889320 -12.11963146 -12.16036972 -12.20110797
#> 
#> $mlests
#> [1] 5.180922e-19 1.850448e+00
#>